csharp-basics/exercises/ClassesAndObjects/Account/Account.cs
namespace Account
{
    public class Account
    {
        private string _name;
        private double _money;

        public Account(string v1, double v2)
        {
            _name = v1;
            _money = v2;
        }

        public double Withdrawal(double i)
        {
             _money -= i;
             return _money;
        }

        public double Deposit(double i)
        {
            _money += i;
            return _money;
        }

        public double Balance()
        {
            return _money;
        }

        public override string ToString()
        {
            return $"{_name}: {_money.ToString("0.0")}";
        }

        public static void Transfer(Account from, Account to, double howMuch)
        {
            to.Deposit(from.Withdrawal(howMuch));
        }
    }
}
